Reorganise server dir struct
