+++ server/src/styles/altair.py
